Add client pooling, timeouts and retries
This is a breaking change to the API which simplifies things for the
API consumer and adds features to help when iRODS operations time out
because of network or server issues.

The constructors of Collection and DataObject no longer take a Baton
client as their first argument. Instead, Collection and DataObject
reference a pool of clients which they draw on to perform
operations. This means that if a client encounters problems, it can be
returned to the pool to be restarted. If no pool is specified on
construction, a default, threadsafe pool of 4 clients is used, shared
between all Collection and DataObjects.

All remote operations now support a timeout and a number of
tries. These default to no timeout and 1 try, to retain the original
behaviour.