Fix subprocess communication to avoid restarting the client
An earlier change introduced the use of Popen.communicate to send and
recieve from the baton-do subprocess without understanding that
Popen.communicate will terminate the subprocess afterwards. This led
to subprocesses being restartd with every request.

The motivation for using Popen.communicate was to avoid deadlocks "due
to any of the other OS pipe buffers filling up and blocking the child
process".