Fix blocking on STDERR
When baton-do wrote to STDERR, it would eventually fill the buffer and
block because nothing was reading.

Add a reader daemon thread for STDERR.

Remove setting of a "blocked" flag when a client times out. Instead,
stop the client and raise an Exception. The client will return to the
pool and be restarted automatically if used again.

Dispense with the unnecessary self._mutex. Baton instances are not
threadsafe.

Run baton-do with --silent and --no-error because we don't really want
these error messages anyway. The same information is provided in the
JSON output on STDOUT.