Bug: do not cache rods_type, but set whewn first known
Caching the rods_type had the side effect of making paths whose type
was not knowable by the constructor, permanently unknown. E.g. a for a
DataObject whose path does not yet exist. This change avoids blindly
caching None as a type and checks the type as required, until it is
known, from which point it no longer changes.

Remove an unnecessary f-string.
Improve a docstring.