Add the option to capture exceptions from Collection::put
Add a keyword argument to Collection::put that causes the generator to
return any Exception raised when putting an item. By returning the
Excpetion rather than raising it, the generator can continue to do as
much work as possible before completing, rather than terminating
immediately. This is useful for bulk uploads where we would otherwise
have to restart the whole traversal for each error. The default
behaviour remains the same as it was; an exception is raised as soon
as the first error is encountered.